fix(product): satisfy frontend structure guard
